+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 - Fixed changing to another hardware will clear the filter component search
 - Added "Oscilloscope" tab
 - Added oscilloscope sync when viewing oscilloscope images in component info popup
+- Added oscilloscope labels in the top-right corner for the large image in component info popup
 - Changed minor UI and text changes